Enhanced safety of interlocked writer

+++ src/nwave/interlocked/writer.py
@@ -7,10 +7,11 @@ import tempfile
 
 
 class Writer:
-    def __init__(self, file: str | PathLike):
+    def __init__(self, file: str | PathLike, overwrite: bool = False):
         if os.path.isdir(file):
             raise ValueError(f"{file} cannot be a directory")
         self.file = file
+        self.overwrite = overwrite
         self.temp = tempfile.NamedTemporaryFile(
             delete=False,
             dir=os.path.dirname(self.file),
@@ -24,8 +25,16 @@ class Writer:
     def __exit__(self, exc_type, exc_val, exc_tb):
         # Close temp file
         self.temp.close()
-        # If target file exists, delete it
-        if os.path.isfile(self.file):
-            os.remove(self.file)
-        # Copy temp file to target (rename)
-        os.rename(self.temp.name, self.file)
+        try:
+            # If target file exists
+            if os.path.isfile(self.file):
+                # Raise error if overwrite is False
+                if not self.overwrite:
+                    raise FileExistsError(f"File {self.file} already exists")
+                os.remove(self.file)
+            # Copy temp file to target (rename)
+            os.rename(self.temp.name, self.file)
+        finally:
+            # Delete temp file
+            if os.path.exists(self.temp.name):
+                os.remove(self.temp.name)
